+++ docs/CHANGE_LOG.md
@@ -4,6 +4,7 @@
 
 #### **Start() > Options:**
 - Added option "showNodeNameInTitle" (defaults to false), which states if the node name should be shown in the title (if more than one node type is being used).
+- Renamed option "cssPropertiesText" to "cssText", and changed the default value to "CSS".
 
 <br>